# backend/src/routes/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"IUser": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"age": {
						"type": "number",
						"format": "double"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"IPagination_IUser-Array_": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/IUser"
						},
						"type": "array"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"size": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"page",
					"limit",
					"size"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IResponse_IPagination_IUser-Array__": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/IPagination_IUser-Array_"
					},
					"status": {
						"type": "string"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"status",
					"code",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IResponse_IUser_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/IUser"
					},
					"status": {
						"type": "string"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"status",
					"code",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Model": {
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"IValidationError": {
				"properties": {
					"target": {
						"$ref": "#/components/schemas/Model"
					},
					"property": {
						"type": "string"
					},
					"value": {},
					"constraints": {
						"properties": {},
						"additionalProperties": {
							"type": "string"
						},
						"type": "object"
					},
					"children": {
						"items": {
							"$ref": "#/components/schemas/IValidationError"
						},
						"type": "array"
					},
					"contexts": {
						"properties": {},
						"additionalProperties": {},
						"type": "object"
					}
				},
				"required": [
					"property",
					"constraints",
					"children"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IResponse_IUser-or-IValidationError-Array_": {
				"properties": {
					"data": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/IUser"
							},
							{
								"items": {
									"$ref": "#/components/schemas/IValidationError"
								},
								"type": "array"
							}
						]
					},
					"status": {
						"type": "string"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"status",
					"code",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/users": {
			"get": {
				"operationId": "List",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_IPagination_IUser-Array__"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "size",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_IUser-or-IValidationError-Array_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUser"
							}
						}
					}
				}
			}
		},
		"/users/{id}": {
			"get": {
				"operationId": "Item",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_IUser_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_IUser-or-IValidationError-Array_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUser"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}